requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
numpy
Pillow
keras
tensorflow
